refactor(frontend): extract useOccupancyHierarchyFilter hook from GanttChart
Bundle the occupancy (Standort/Parzelle/Beet) tree concern into a
useOccupancyHierarchyFilter hook: the location/field/only-occupied filters,
the persisted expand/collapse state with its one-time default expansion, the
level toggle, and the derived hierarchy nodes, field options, active filter
count, and reset. The page passes the loaded data + displayYear +
activeProjectId and consumes the returned values in task-group building and
the filter UI. Search text stays in the page (shared with the seedling view).
GanttChart drops the now-unused useExpandedState, useHierarchyLevelToggle,
buildFieldOccupancyHierarchy, and threshold imports. No behavior change.
